spsc-ring: assert power-of-two nslots; widen overflow counter to u64
spsc_ring_init did not validate nslots, but the three consumers
use mask = nslots - 1 + & mask as the slot modulo, which is only
correct for powers of two.  Every current caller passes a
power-of-two literal; the assert keeps a future caller from
silently corrupting the ring.  nslots is not in scope at init
time, so the assert lives at the top of each function that
consumes it (try_enqueue, drain, overwrite_enqueue).

r->overflow was uint32_t, wrapping at ~1M edges/sec in seconds
under sustained backpressure.  Widen to uint64_t through the
struct, drain function, and its callers' overflow_out locals.

+++ edgepair-ring.c
@@ -224,7 +224,7 @@ static void apply_slot(const void *p, void *ctx __unused__)
 
 unsigned int edgepair_ring_drain(struct edgepair_ring *ring)
 {
-	uint32_t overflow = 0;
+	uint64_t overflow = 0;
 	uint32_t processed;
 
 	if (ring == NULL)
